Move out components to their own folders
