+++ src/App.tsx
@@ -22,7 +22,7 @@ import '@ionic/react/css/display.css';
 /* Theme variables */
 import './theme/variables.css';
 import MyReminders from './pages/MyReminders';
-import CountryElections from './components/CountryElections';
+import CountryElections from './components/CountryElections/CountryElections';
 
 setupIonicReact();
