+++ src/components/ElectionCard/ElectionCard.tsx
