+++ src/components/IntroductionTextCard/IntroductionTextCard.tsx
