+++ src/utilFunctions/countryElections/onIonChange.ts
@@ -1,0 +1,14 @@
+function onIonChange(userEnteredValue: String | null | undefined, setFilterTerm: Function, setDummyElectionDataResults: Function, filterFields: Object) {
+    /*
+    setFilterTerm(userEnteredValue!);
+    if (userEnteredValue! === '') {
+        setDummyElectionDataResults(dummyElectionData);
+        return;
+    }
+    if (filterTypeTerm === filterFields.NAME) {
+        const resultsFilteredByName = dummyElectionDataResults.filter(({ electionName }) => electionName === userEnteredValue);
+        setDummyElectionDataResults(resultsFilteredByName);
+    }
+    */
+
+}
